--- packages/java/tests/spring/react-i18n/frontend/App.tsx
@@ -1,6 +1,0 @@
-import router from 'Frontend/routes.js';
-import { RouterProvider } from 'react-router';
-
-export default function App() {
-  return <RouterProvider router={router} />;
-}
