--- packages/java/tests/spring/react-i18n/frontend/index.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>i18n</title>
-    <style>
-      body {
-        margin: 0;
-        width: 100vw;
-        height: 100vh;
-      }
-
-      #outlet {
-        height: 100%;
-      }
-    </style>
-    <!-- index.ts is included here automatically (either by the dev server or during the build) -->
-  </head>
-
-  <body>
-    <div id="outlet"></div>
-  </body>
-</html>
