--- packages/java/tests/spring/react-i18n/frontend/routes.tsx
@@ -1,18 +1,0 @@
-import { createBrowserRouter } from 'react-router';
-import MainLayout from './MainLayout.js';
-import BasicI18NView from './_views/BasicI18NView';
-export const routes = [
-  {
-    path: '',
-    element: <MainLayout />,
-    children: [
-      {
-        path: '/basic-i18n',
-        element: <BasicI18NView />,
-      },
-    ],
-  },
-];
-
-const router = createBrowserRouter([...routes]);
-export default router;
