+++ packages/java/tests/spring/react-i18n/src/main/frontend/themes/react-i18n-test/theme.json
