+++ packages/java/tests/spring/react-i18n/src/main/frontend/views/detail.ts
@@ -1,0 +1,3 @@
+export type Detail = {
+  description?: string;
+};
